+++ Divider.h
@@ -5,7 +5,7 @@
 
 
 using namespace std;
-extern int constructValueToRegister(int value);
+extern int constructValueToRegister(long long value);
 extern void	undefinedVariableError(string varName);
 extern int	checkInitialization(string varName);
 extern map<string, MemoryItem*> memoryMap;
@@ -15,14 +15,14 @@ extern vector<AsmInstruction*> asmInstrunctions;
 class Divider
 {
 	public:
-    static int prepare(int, int);
-    static int prepare(int, string);
-    static int prepare(string, int);
+    static int prepare(long long, long long);
+    static int prepare(long long, string);
+    static int prepare(string, long long);
     static int prepare(string, string);
 		static void doTheJob();
 };
 
-int Divider::prepare(int a, int b)
+int Divider::prepare(long long a, long long b)
 {
   constructValueToRegister(a);
   asmInstrunctions.push_back(new AsmInstruction("STORE", 9));
@@ -31,7 +31,7 @@ int Divider::prepare(int a, int b)
 	return 0;
 }
 
-int Divider::prepare(int a, string name)
+int Divider::prepare(long long a, string name)
 {
 	map<string, MemoryItem*>::iterator it = memoryMap.find(name);
 	if(it == memoryMap.end())
@@ -51,7 +51,7 @@ int Divider::prepare(int a, string name)
 	}
 }
 
-int Divider::prepare(string name, int a)
+int Divider::prepare(string name, long long a)
 {
 	map<string, MemoryItem*>::iterator it = memoryMap.find(name);
 	if(it == memoryMap.end())
