+++ Substractor.h
@@ -4,7 +4,7 @@
 #include <map>
 
 using namespace std;
-extern int constructValueToRegister(int value);
+extern int constructValueToRegister(long long value);
 extern void	undefinedVariableError(string varName);
 extern int	checkInitialization(string varName);
 extern map<string, MemoryItem*> memoryMap;
